requirements.txt
fastapi~=0.63.0
uvicorn~=0.13.4
pydantic~=1.8.1
aiofiles~=0.7.0
motor~=2.3.1
pymongo~=3.11.3
python-dotenv~=0.15.0
requests~=2.25.1
PyJWT~=2.0.1
email-validator~=1.1.2
passlib~=1.7.4
aiofiles~=0.7.0
boto3~=1.16.63
python-magic~=0.4.24
pixellib~=0.7.1
starlette~=0.13.6